test: migrate config package from `jest` to `ava`, part 3
